refactor: remove cluster overview utils re-export
