+++ src/containers/Cluster/ClusterOverview/__test__/utils.test.ts
@@ -1,4 +1,7 @@
-import {calculateBaseDiagramValues, getDiagramValues} from '../utils';
+import {
+    calculateBaseDiagramValues,
+    getDiagramValues,
+} from '../../../../utils/metrics/getDiagramValues';
 
 describe('ClusterOverview utils', () => {
     test('formats metric percentages with one decimal only below one percent', () => {
