+++ src/containers/Cluster/ClusterOverview/components/ClusterMetricsStorage.tsx
@@ -1,9 +1,9 @@
 import {DoughnutMetrics} from '../../../../components/DoughnutMetrics/DoughnutMetrics';
 import {normalizeMediaType} from '../../../../utils/disks/normalizeMediaType';
 import {formatStorageLegend} from '../../../../utils/metrics/formatMetricLegend';
+import {getDiagramValues} from '../../../../utils/metrics/getDiagramValues';
 import i18n from '../../i18n';
 import type {ClusterMetricsCommonProps} from '../shared';
-import {getDiagramValues} from '../utils';
 
 import {ClusterMetricsCardContent} from './ClusterMetricsCard';
