--- src/containers/Cluster/ClusterOverview/utils.tsx
@@ -1,4 +1,0 @@
-export {
-    calculateBaseDiagramValues,
-    getDiagramValues,
-} from '../../../utils/metrics/getDiagramValues';
